# packages/pydantic_lambda_handler/pydantic_lambda_handler-1.0.0.tar.gz/pydantic_lambda_handler-1.0.0/open_api_spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "PydanticLambdaHandler",
        "version": "0.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/with_headers": {
            "get": {
                "parameters": [
                    {
                        "name": "user_agent",
                        "in": "header",
                        "schema": {
                            "title": "User Agent",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/with_headers_alias": {
            "get": {
                "parameters": [
                    {
                        "name": "UserId",
                        "in": "header",
                        "schema": {
                            "title": "Userid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
...
}